rename batch collection script

+++ tools/separate-2/separate-collection.sh
@@ -1,4 +1,9 @@
-localDB=/Users/benoit/Scripts/custom/free-idea/sources/metro
+if [[ $# -eq 0 ]] ; then
+    echo 'no argument'
+    exit 1
+fi
+
+localDB=$1
 rm -rf cache*
 
 for step in `find ${1:-$localDB} -iname "*.JP*G" -exec ls -l {} \; | awk '{ print $5,"",$9 }'|sort -n`
@@ -10,7 +15,7 @@ do
 	originPath=$(dirname $1)
 	collection=$(basename $1)
 
-	resultdir="/Users/benoit/Scripts/custom/free-idea/sources/"$collection"-result/"$scanid"/"
+	resultdir="/Users/benoit/Scripts/free-idea/free-idea-code/_fragments/"$collection"-result/"$scanid"/"
 
 	if [ ! -d "$resultdir" ]; then
